Add files via upload
The BMI Calculator App could indeed be a software operation that avoids further homemade hours that bears must spend in personally calculating and finding the BMI for a named person at one click. This operation keeps both the quality in it ie American standard and Indian standard too. This app gives us all the knowledge in both the norms which is not given in being app. The main compass is to want care of the health. The BMI App gives us all the data ie it gives suggestions for our health and tells us what should we eat and what to avoid. After we enter the peak and weight we get all the knowledge ie are we fat or light etc...